requirements_base.txt
absl-py==1.0.0
aiohttp==3.8.3
aioitertools==0.7.1
aiosignal==1.2.0
alabaster==0.7.12
altgraph==0.17.4
anaconda-catalogs==0.2.0
anaconda-client==1.11.2
anaconda-navigator==2.4.2
anaconda-project==0.11.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.9.0
astropy==5.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
auto-py-to-exe==2.40.0
Automat==20.2.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bitarray==2.5.1
bkcharts==0.2
black==0.0
bleach==5.0.0
bokeh==2.4.3
boltons==23.0.0
boto3==1.24.28
botocore==1.27.59
bottle==0.12.25
bottle-websocket==0.2.9
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.0.0
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.2.1
clyent==1.2.2
colorama==0.4.6
colorcet==3.0.1
conda==23.7.2
conda-build==3.24.0
conda-content-trust==0+unknown
conda-libmamba-solver==23.5.0
conda-pack==0.6.0
conda-package-handling==2.1.0
conda_package_streaming==0.8.0
conda-repo-cli==1.0.4
conda-token==0.4.0
conda-verify==3.4.2
constantly==15.1.0
cookiecutter==1.7.3
cryptography==41.0.2
cssselect==1.1.0
cycler==0.11.0
Cython==3.0.0
cytoolz==0.12.0
daal4py==2023.1.1
dask==2022.2.1
datashader==0.15.1
datashape==0.5.4
dbus-python==1.3.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
distributed==2022.2.1
docutils==0.18.1
Eel==0.16.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
flake8==4.0.1
Flask==2.2.2
flatbuffers==2.0
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gast==0.5.3
gensim==4.3.0
gevent==23.9.1
gevent-websocket==0.10.1
glob2==0.7
gmpy2==2.1.2
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.44.0
h5py==3.6.0
HeapDict==1.0.1
holoviews==1.17.0
httplib2==0.22.0
hvplot==0.8.4
hyperlink==21.0.0
idna==3.4
imagecodecs==2021.8.26
imageio==2.31.1
imagesize==1.4.1
importlib-metadata==6.0.0
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
intake==0.6.2
intervaltree==3.1.0
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.9.3
itemadapter==0.3.0
itemloaders==1.0.4
itsdangerous==2.0.1
jaraco.classes==3.2.1
jdcal==1.4.1
jedi==0.18.1
jeepney==0.7.1
jellyfish==0.9.0
Jinja2==3.1.1
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.2.0
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab==2.3.1
jupyterlab-pygments==0.2.0
jupyterlab_server==2.22.1
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.13.1
kiwisolver==1.4.4
lazy-object-proxy==1.6.0
libarchive-c==2.9
libclang==13.0.0
libmambapy==1.4.1
llvmlite==0.40.0
locket==1.0.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mccabe==0.7.0
mflt-build-id==1.0.0
mistune==0.8.4
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==8.12.0
mpmath==1.2.1
msgpack==1.0.3
multidict==6.0.2
multipledispatch==0.6.0
munkres==1.1.4
mypy-extensions==0.4.3
navigator-updater==0.4.0
nbclient==0.5.13
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8.4
nltk==3.8.1
nose==1.3.7
notebook==6.4.10
numba==0.57.0
numexpr==2.8.4
numpy==1.22.3
numpydoc==1.5.0
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
panel==0.14.3
param==1.13.0
parsel==1.6.0
parso==0.8.3
partd==1.2.0
pathlib==1.0.1
pathspec==0.10.3
patsy==0.5.3
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.2.1
pkginfo==1.9.6
platformdirs==2.5.1
plotly==4.14.3
pluggy==1.0.0
ply==3.11
pooch==1.4.0
poyo==0.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
Protego==0.1.16
protobuf==4.24.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycosat==0.6.4
pycparser==2.21
pyct==0.5.0
pycurl==7.45.2
PyDispatcher==2.0.5
pydocstyle==6.3.0
pyelftools==0.28
pyerfa==2.0.0
pyflakes==2.4.0
Pygments==2.11.2
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
PyJWT==2.4.0
pylint==2.12.2
pyls-spyder==0.4.0
pyodbc==4.0.34
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.5.0
python-slugify==5.0.2
pytz==2022.7
pyviz-comms==2.3.0
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.2.2
qtconsole==5.3.0
QtPy==2.0.1
queuelib==1.5.0
regex==2022.7.9
requests==2.29.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
responses==0.13.3
retrying==1.3.3
rope==0.22.0
rsa==4.8
Rtree==1.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.17.21
s3transfer==0.6.0
sacremoses==0.0.43
scikit-image==0.19.3
scikit-learn==1.2.2
scikit-learn-intelex==20230426.111436
scipy==1.10.1
Scrapy==2.8.0
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
service-identity==18.1.0
setuptools==67.8.0
sip==6.6.2
six==1.16.0
smart-open==5.2.1
snowballstemmer==2.2.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.2
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.3.3
spyder-kernels==2.3.3
SQLAlchemy==1.4.27
stack-data==0.2.0
statsmodels==0.14.0
sympy==1.11.1
tables==3.8.0
tabulate==0.8.10
TBB==0.2
tblib==1.7.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
textdistance==4.2.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss==0.4
tinycss2==1.1.1
tldextract==3.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.11.0
tornado==6.1
tqdm==4.65.0
traitlets==5.1.1
Twisted==22.10.0
typing_extensions==4.1.1
ujson==5.4.0
Unidecode==1.2.0
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.14.1
w3lib==1.21.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
whatthepatch==1.0.2
wheel==0.37.1
whichcraft==0.6.1
widgetsnbextension==3.6.0
wrapt==1.14.0
wurlitzer==3.0.2
xarray==2023.6.0
xlrd==2.0.1
XlsxWriter==3.1.1
yapf==0.31.0
yarl==1.8.1
zict==2.1.0
zipp==3.11.0
zope.event==5.0
zope.interface==5.4.0
zstandard==0.19.0
